+++ medical-record/medical-record/src/main/java/com/ivanovp/medical_record/service/UserService.java
@@ -1,0 +1,12 @@
+package com.ivanovp.medical_record.service;
+
+import com.ivanovp.medical_record.dto.response.UserResponseDTO;
+
+import java.util.List;
+
+public interface UserService {
+
+    List<UserResponseDTO> getAllUsers();
+
+    void deleteUser(Long id);
+}